+++ core/src/main/java/com/cryptape/cita/protocol/core/methods/response/CITASha3.java
@@ -3,7 +3,7 @@ package com.cryptape.cita.protocol.core.methods.response;
 import com.cryptape.cita.protocol.core.Response;
 
 /**
- * web3_sha3.
+ * CITA_sha3.
  */
 public class CITASha3 extends Response<String> {
 }