+++ crypto/src/main/java/com/cryptape/cita/crypto/SecureRandomUtils.java
@@ -6,7 +6,6 @@ import java.security.SecureRandom;
  * Utility class for working with SecureRandom implementation.
  *
  * <p>This is to address issues with SecureRandom on Android. For more information, refer to the
- * following <a href="https://github.com/web3j/web3j/issues/146">issue</a>.
  */
 final class SecureRandomUtils {