+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 All notable changes to this project will be documented in this file.
 
-# [v0.20.3](https://github.com/cryptape/cita-sdk-swift/compare/v0.20...v0.20.3) (2019-01-21)
+# [v0.21](https://github.com/cryptape/cita-sdk-java/compare/v0.20.3...v0.21) (2019-02-18)
+
+### BREAKING CHANGES
+
+* Update Crypto field of protobuf. 
+* Add field from to transaction response.
+
+# [v0.20.3](https://github.com/cryptape/cita-sdk-java/compare/v0.20...v0.20.3) (2019-01-21)
 
 ### BREAKING CHANGES